decoupled login api from websocket server
