+++ lib/authorize.js
@@ -120,4 +120,4 @@ function authorize(options, onConnection) {
     }
 }
 
-exports.authorize = authorize;
+module.exports = authorize;
