+++ lib/index.js
@@ -1,8 +1,8 @@
 var assert = require('assert');
-var socketioSec = require('./authorize');
+var authorize = require('./authorize');
 var jwt = require('jsonwebtoken');
 var socketIo = require('socket.io');
-
+var api = require('./api');
 /**
  * 
  * 
@@ -19,7 +19,8 @@ var socketIo = require('socket.io');
  * + If not, the web socket server will send back "authenticated""
  * 
  * 
- * ----------------------------------------------------------------------- * Challenge 1:
+ * ----------------------------------------------------------------------- 
+ * Challenge 1:
  * ------------
  * if the token has expired and the connection has never been disconnected, the client app will still use the connection.
  * 
@@ -88,18 +89,23 @@ var socketIo = require('socket.io');
 * 
 */
 
-module.exports.server = function(server, app, options) {
-    
-    assert.notStrictEqual(options.claim,undefined);
-    assert.notStrictEqual(options.findUserByCredentials,undefined);
+
+/**
+ * 
+ * create a socketio server with middleware to handle token based connection, reconnection
+ * 
+ * return a new instance of the socketIo server 
+ * 
+ */
+function socketServe(server, app, options) {
+
+    assert.notStrictEqual(options.claim, undefined);
+    assert.notStrictEqual(options.findUserByCredentials, undefined);
 
     var io = socketIo.listen(server);
 
     if (!options.refresh) {
         options.refresh = generateSessionToken;
-    }
-    if (!options.authorization) {
-        options.authorization = generateAuthorizationCode;
     }
 
     if (!options.claim) {
@@ -107,65 +113,38 @@ module.exports.server = function(server, app, options) {
     }
 
     io.sockets
-        .on('connection', socketioSec.authorize(options));
-
-    app.post('/login', handleLoginRequest);
-
-    app.post('/register', handleRegisterRequest);
+        .on('connection', authorize(options));
 
     return io;
 
     ////////////////////////////////////////
 
-    function handleLoginRequest(req, res) {
-        console.log('check credentials for ' + JSON.stringify(req.body) + ', then send back auth code...');
-        var user = options.findUserByCredentials(req.body)
-            .then(function(user) {
-                    sendAuthorizationResponse(res, user);
-            })
-            .catch(function(error) {
-                res.status(401).send({ code: "USER_INVALID" });
-            });
-    }
-
-    function handleRegisterRequest(req, res) {
-        console.log('Registering ' + JSON.stringify(req.body) + ', then send back auth code...');
-
-        options.register(req.body)
-            .then(function(user) {
-                sendAuthorizationResponse(res, user);
-            })
-            .catch(function(err) {
-                //  Bad Request :The server cannot or will not process the request due to an apparent client error (e.g., malformed request syntax, invalid request message framing, or deceptive request routing).
-                res.status(400).send({ code: err });
-            });
-    }
-
-    function sendAuthorizationResponse(res, user) {
-        var token = options.authorization(options.claim(user));
-        if (options.appUrl) {
-            res.status(200).send(options.appUrl(token));
-        } else {
-            res.json({ token: token });
-        }
-    }
-
-
-    // function checkPassword(userPassword, password) {
-    //     // @TODO: userPassword should be encrypted 
-    //     // passwork must be encrypted before comparing with the current user password
-    //     return password == "123";
-    // }
-
-    function generateAuthorizationCode(payload) {
-        return jwt.sign(payload, options.secret, { 
-            expiresIn: 10
-        });
-    }
-
     function generateSessionToken(payload) {
         payload.dur = 5 * 60;
         return jwt.sign(payload, this.secret, { expiresIn: payload.dur });
     }
+};
+
+/**
+ * add to express app the login and registration api request handling.
+ */
+function apiServe(app, options) {
+    return api(app, options);
+}
+
+/**
+ * sugar function that sets up a secure socket server and provide api on the same server
+ * 
+ * return a new instance of the socketIo server 
+ */
+function infrastructure(server, app, options) {
+    apiServe(app, options);
+    return socketServe(server, app, options);
+};
+
+module.exports = {
+    apiServe: apiServe,
+    socketServe: socketServe,
+    infrastructure: infrastructure
 };
